main-docker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP 클라이언트
httpx==0.25.2

# 인증 및 보안
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Firebase
firebase-admin==6.2.0
google-cloud-firestore==2.13.1

# 기타 유틸리티
python-json-logger==2.0.7 